feat(weather): config url 수정
